refactor: Move to registration package
